new hook for easier panel event usage
