+++ app/packages/core/src/plugins/SchemaIO/components/NativeModelEvaluationView/utils.ts
@@ -1,29 +1,7 @@
-import { usePanelEvent } from "@fiftyone/operators";
-import { usePanelId } from "@fiftyone/spaces";
 import { formatValueAsNumber } from "@fiftyone/utilities";
 import { capitalize } from "lodash";
-import { useCallback } from "react";
 import { atom } from "recoil";
 import { NONE_CLASS } from "./constants";
-
-export function useTriggerEvent() {
-  const panelId = usePanelId();
-  const handleEvent = usePanelEvent();
-
-  const triggerEvent = useCallback(
-    (event: string, params?: any, prompt?: boolean, callback?: any) => {
-      handleEvent(panelId, {
-        operator: event,
-        params,
-        prompt,
-        callback,
-      });
-    },
-    [handleEvent, panelId]
-  );
-
-  return triggerEvent;
-}
 
 export function getNumericDifference(
   value,
@@ -40,6 +18,7 @@ export function getNumericDifference(
     }
     return formatValueAsNumber(difference, fractionDigits);
   }
+  return NaN;
 }
 
 export function getMatrix(
