+++ app/packages/operators/src/index.ts
@@ -28,4 +28,7 @@ export {
   useOperatorPlacements,
 } from "./state";
 export * as types from "./types";
-export { default as usePanelEvent } from "./usePanelEvent";
+export {
+  default as usePanelEvent,
+  useTriggerPanelEvent,
+} from "./usePanelEvent";
